integration test appsettings now override default appsettings.json
