+++ ACI.Products.Test/ACI.Products.Test.csproj
@@ -8,7 +8,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <FrameworkReference Include="Microsoft.AspNetCore.App" />
         <PackageReference Include="Bogus" Version="34.0.1" />
         <PackageReference Include="FluentAssertions" Version="6.5.1" />
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="6.0.3" />
@@ -27,6 +26,12 @@
 
     <ItemGroup>
       <ProjectReference Include="..\ACI.Products\ACI.Products.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Content Update="appsettings.json">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </Content>
     </ItemGroup>
 
 </Project>
