--- ACI.Products.Test/Integration/CustomWebApplicationFactory.cs
@@ -1,52 +1,0 @@
-using System;
-using System.Linq;
-using ACI.Products.Data;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-
-namespace ACI.Products.Test.Integration;
-
-public class CustomWebApplicationFactory<TStartup>
-    : WebApplicationFactory<TStartup>
-    where TStartup : class
-{
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        builder.ConfigureTestServices(services =>
-        {
-            var descriptor = services.SingleOrDefault(
-                d => d.ServiceType ==
-                     typeof(DbContextOptions<ProductContext>));
-
-            services.Remove(descriptor);
-
-            services.AddDbContext<ProductContext>(options =>
-            {
-                options.UseInMemoryDatabase("InMemoryDbForTesting");
-            });
-
-            var sp = services.BuildServiceProvider();
-
-            using var scope = sp.CreateScope();
-            var scopedServices = scope.ServiceProvider;
-            var db = scopedServices.GetRequiredService<ProductContext>();
-            var logger = scopedServices
-                .GetRequiredService<ILogger<CustomWebApplicationFactory<TStartup>>>();
-
-            db.Database.EnsureCreated();
-
-            try
-            {
-                DbSetup.InitializeForTests(db);
-            }
-            catch (Exception ex)
-            {
-                logger.LogError(ex, "An error occurred seeding the database with test messages. Error: {Message}", ex.Message);
-            }
-        });
-    }
-}
